WIP, fixing inspectors profit collect & adding long tail inspector & better atomic arbitrage

+++ crates/brontes-types/src/db/dex.rs
@@ -70,7 +70,7 @@ impl DexQuotes {
         }
 
         loop {
-            if let Some(price) = self.get_price(pair, tx) {
+            if let Some(price) = self.get_price(pair, tx as u64) {
                 return Some(price.clone());
             }
             if tx == 0 {
@@ -90,8 +90,8 @@ impl DexQuotes {
             .unwrap_or(false)
     }
 
-    fn get_price(&self, pair: Pair, tx: usize) -> Option<&DexPrices> {
-        self.0.get(tx)?.as_ref()?.get(&pair)
+    fn get_price(&self, pair: Pair, tx: u64) -> Option<&DexPrices> {
+        self.0.get(tx as usize)?.as_ref()?.get(&pair)
     }
 }
